+++ NEWS.md
@@ -1,2 +1,6 @@
 # cmtools v0.0.1
+* Added functions to calculate the sandwich estimator for the covariance matrix
+when the model has been solved using BGW. 
+* Added function to add the scores to a BGW object
+* Added broom generics for BGW objects
 * Added function for the Poe et al. (2005) test